+++ CHANGELOG.md
@@ -7,3 +7,6 @@ Features:
 - Define the `ObjectSerializer` interface and its `JsonObjectSerializer` implementation.
 - Implement the `JsonSerializableBigInt` decorator.
 - Define the `Event` and `EventPublisher` interfaces.
+- Define the `EventTransaction` interface and implement the `BufferEventTransaction`.
+- Implement the `Transaction` class.
+- Implement the base `TransactionRunner`.